fix handling of retries in RavelloClient
The logic which handles "retries" in RavelloClient._request() is broken.
If you want to disable retries altogether, you must set retries=1;
setting retries=0 causes a failure.

To be explicit, this means that retries=2 actually retries once and
retries=3 only retries twice. This is unintuitive.

This commit fixes all that (so that 0 retries is 0 and 3 is really 3) by
simply changing the sign in "while retries < self.retries" to "<=".